Move styling extension to the end of `WooBasicTableViewCell`.
