--- WooCommerce/Classes/Extensions/WooBasicTableViewCell+Styling.swift
@@ -1,18 +1,0 @@
-import UIKit
-
-extension WooBasicTableViewCell {
-    func applyListSelectorStyle() {
-        bodyLabel.applyBodyStyle()
-        bodyLabelTopMarginConstraint.constant = 0
-    }
-
-    func applyPlainTextStyle() {
-        bodyLabel.applyBodyStyle()
-        bodyLabelTopMarginConstraint.constant = 8
-    }
-
-    func applyActionableStyle() {
-        bodyLabel.applyActionableStyle()
-        bodyLabelTopMarginConstraint.constant = 8
-    }
-}
