+++ WooCommerce/Classes/ViewRelated/ReusableViews/WooBasicTableViewCell.swift
@@ -65,3 +65,21 @@ class WooBasicTableViewCell: UITableViewCell {
         accessoryView = accessoryImageView
     }
 }
+
+extension WooBasicTableViewCell {
+    func applyListSelectorStyle() {
+        bodyLabel.applyBodyStyle()
+        bodyLabelTopMarginConstraint.constant = 0
+    }
+
+    func applyPlainTextStyle() {
+        bodyLabel.applyBodyStyle()
+        bodyLabelTopMarginConstraint.constant = 8
+    }
+
+    func applyActionableStyle() {
+        bodyLabel.applyActionableStyle()
+        bodyLabelTopMarginConstraint.constant = 8
+    }
+}
+
